# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Home-Navi - Care Facility Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weights:
    get:
      operationId: listWeights
      tags: [weights]
      summary: List weight records
      parameters:
        - name: resident_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: this_month
          in: query
          schema:
            type: boolean
            description: If true, only return this month's records
      responses:
        "200":
          description: List of weight records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Weight"
    post:
      operationId: createWeight
      tags: [weights]
      summary: Record weight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeightBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weight"
  /weights/{id}:
    patch:
      operationId: updateWeight
      tags: [weights]
      summary: Update a weight record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWeightBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weight"
    delete:
      operationId: deleteWeight
      tags: [weights]
      summary: Delete a weight record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Weight:
      type: object
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: ["string", "null"]
        recordedAt:
          type: string
        weightKg:
          type: number
        photoUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - residentId
        - recordedAt
        - weightKg
        - createdAt
    CreateWeightBody:
      type: object
      properties:
        residentId:
          type: integer
        recordedAt:
          type: string
        weightKg:
          type: number
        photoUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - residentId
        - recordedAt
        - weightKg
    UpdateWeightBody:
      type: object
      properties:
        weightKg:
          type: ["number", "null"]
        photoUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...